Reorganized scripts, remove unnecessary tests and plots.
