+++ UneccessaryPlots/test4_single_robot.py
@@ -7,34 +7,38 @@ This test file is dependent on vrep.
 """
 
 from scene import Scene
-from sceneplot import ScenePlot
+from useful.sceneplot import ScenePlot
 # from robot import Robot
 import numpy as np
 # from data import Data
 
-def generateData():
+def generateData(**kwargs):
     sc = Scene(recordData = True)
     sp = ScenePlot(sc)
     try:
-        dynamics = 11
+        #dynamics = 20
+        arg2 = np.float32([.5, .5])
+        for name, value in kwargs.items():
+            if name == "dynamics":
+                dynamics = value
+            elif name == "arg2":
+                arg2 = value
         sc.dynamics = dynamics
-        sc.addRobot(np.float32([[-2, 0, 0], [0, 2/2, 0]]))
-        sc.addRobot(np.float32([[1, 3, 0], [1.732/2, -1/2, 0]]))
-        sc.addRobot(np.float32([[0, 0, 0], [-1.732/2, -1/2, 0]]))
+        sc.addRobot(np.float32([[0, 0, 0], [0, 2/2, 0]]), arg2)
         
         # No leader
-        sc.setADjMatrix(np.uint8([[0, 1, 1], [1, 0, 1], [1, 1, 0]]))
+        sc.setADjMatrix(np.uint8([[0]]))
         # Set robot 0 as the leader.
         # sc.setADjMatrix(np.uint8([[0, 0, 0], [1, 0, 1], [1, 1, 0]]))
         
         # vrep related
         sc.initVrep()
         # Choose sensor type
-        sc.SENSOR_TYPE = "VPL16" # None, 2d, VPL16, kinect
+        sc.SENSOR_TYPE = "None" # None, 2d, VPL16, kinect
         sc.objectNames = ['Pioneer_p3dx', 'Pioneer_p3dx_leftMotor', 'Pioneer_p3dx_rightMotor']
         
         if sc.SENSOR_TYPE == "None":
-            pass
+            sc.setVrepHandles(0, '')
         elif sc.SENSOR_TYPE == "2d":
             sc.objectNames.append('LaserScanner_2D_front')
             sc.objectNames.append('LaserScanner_2D_rear')
@@ -54,32 +58,24 @@ def generateData():
             sc.setVrepHandles(2, '#1')
         
         #sc.renderScene(waitTime = 3000)
-        tf = 10
-        sc.resetPosition()
+        tf = 30
         sp.plot(3, tf)
         while sc.simulate():
             #sc.renderScene(waitTime = int(sc.dt * 1000))
-            sc.showOccupancyMap(waitTime = int(sc.dt * 1000))
+            #sc.showOccupancyMap(waitTime = int(sc.dt * 1000))
             
             #print("---------------------")
             #print("t = %.3f" % sc.t, "s")
             
             #sp.plot(0, tf)
-            sp.plot(2, tf)
-            #sp.plot(1, tf) 
+            #sp.plot(1, tf)
+            #sp.plot(2, tf) 
             sp.plot(3, tf)
             sp.plot(4, tf)
             sp.plot(5, tf)
             sp.plot(6, tf)
             if sc.t > tf:
                 break
-                
-        
-            #print('robot 0: ', sc.robots[0].xi.x, ', ', sc.robots[0].xi.y, ', ', sc.robots[0].xi.theta)
-            #print('robot 1: ', sc.robots[1].xi.x, ', ', sc.robots[1].xi.y, ', ', sc.robots[1].xi.theta)
-            #print('robot 2: ', sc.robots[2].xi.x, ', ', sc.robots[2].xi.y, ', ', sc.robots[2].xi.theta)
-            #print('y01: ' + str(sc.robots[1].xi.y - sc.robots[0].xi.y))
-            #print('x02: ' + str(sc.robots[2].xi.x - sc.robots[0].xi.x))
         sc.deallocate()
     except KeyboardInterrupt:
         x = input('Quit?(y/n)')
@@ -91,42 +87,13 @@ def generateData():
         sc.deallocate()
         raise
     
-    # check max formation error
-    maxAbsError = 0
-    for key in sc.ydict[2]:
-        absError = abs(sc.ydict[2][key][-1])
-        if absError > maxAbsError:
-            maxAbsError = absError
-    print('maxAbsError = ', maxAbsError)
-    
-    if maxAbsError < 0.5:
-        return sc
-    else:
-        return None
+    return None
 # main
-numRun = 100
-dataList = []
+numRun = 6
 
-
-
-
-for i in range(0, numRun):
+for i in range(5, numRun):
     print('Run #: ', i, '...')
-    # First episode
-    sc = generateData()
-    if sc != None:
-        # if the list is empty
-        if not dataList:
-            for robot in sc.robots:
-                dataList.append(robot.data)
-        else:
-            for j in range(len(sc.robots)):
-                dataList[j].append(sc.robots[j].data)
-        
-
-for j in range(len(sc.robots)):
-    dataList[j].store()
-
+    sc = generateData(dynamics = 20, arg2 = i * np.float32([.3, .3]))
